updated the changes as per review
