+++ routes/index.js
@@ -1,15 +1,12 @@
-const express = require('express');
-const {ensureAuthenticated} = require('../config/auth');
-const router = express.Router()
+const indexController = require('../controller/index-controller')
+const router = require('express').Router()
+const {ensureAuthenticated} = require('../middleware/auth');
+
 
 //Home page
-router.get('/', (req,res)=>
-    res.render('welcome')
-)
+router.get('/', indexController.welcome)
+
 //dashboard
-router.get('/dashboard', ensureAuthenticated,(req,res)=>
-    res.render('dashboard',{
-        name:req.user.name
-    })
-)
+router.get('/dashboard',ensureAuthenticated, indexController.dash)
+
 module.exports = router;
