+++ routes/users.js
@@ -1,105 +1,21 @@
 const express = require('express');
 const router = express.Router()
-const bcrypt = require('bcryptjs')
-const passport = require('passport')
-
-//User model
-const User = require('../models/User')
+const usersController = require('../controller/users-controller')
 
 //Login-page
-router.get('/login', (req,res)=>{
-    res.render('login')
-})
+router.get('/login', usersController.login )
 
 //Register page
-router.get('/register', (req,res)=>{
-    res.render('register')
-})
+router.get('/register', usersController.register)
 
 //Register handle
-router.post('/register', (req,res)=>{
-    const {name, email, password, password2} = req.body;
-    let errors = []
-
-// check required fields
-if(!name || !email || !password || !password2 ){
-    errors.push({msg:'Please provide all fields'})
-}
-
-// check password match
-if(password !== password2){
-    errors.push({msg:'Password dont match'})
-}
-
-//password length
-if(password.length < 6){
-    errors.push({msg:'Password should be min 6 characters'})
-};
-
-if(errors.length>0){
-    res.render('register',{
-        errors,
-        name,
-        email,
-        password,
-        password2
-    })
-}else{
-// validation passed
-User.findOne({email:email})
-.then(user =>{
-    if(user){
-        //user exists
-        errors.push({msg:'Email has already registered'})
-        res.render('register',{
-            errors,
-            name,
-            email,
-            password,
-            password2
-        });
-        }else{
-            const newUser = new User({
-                name,
-                email,
-                password
-
-            });
-         //hash password
-            bcrypt.genSalt(10, (err,salt)=>
-            bcrypt.hash(newUser.password, salt, (err,hash)=>{
-                if(err) throw err;
-                //set password hashed
-                newUser.password = hash;
-                //save user
-                newUser.save()
-                .then(user =>{
-                    req.flash('success_msg', 'You are now registered and can login')
-                    res.redirect('/users/login');
-                })
-                .catch(err => console.log(err))
-            }))
-        }
-
-})
-}
-});
+router.post('/register', usersController.registerHandle);
 
 //Login handle
-router.post('/login',(req,res,next)=>{
-    passport.authenticate('local',{
-        successRedirect:'/dashboard',
-        failureRedirect:'/users/login',
-        failureFlash:true
-    })(req,res,next);
-})
+router.post('/login',usersController.loginHandle)
 
 //logout handle
-router.get('/logout', (req,res)=>{
-    req.logout();
-    req.flash('success_msg', 'You are logged out');
-    res.redirect('/users/login')
-})
+router.get('/logout', usersController.logoutHandle)
 
 
 module.exports = router;
